+++ packages/labs/analyzer/test-files/ts/basic-elements/src/class-with-unsupported-property.ts
@@ -1,0 +1,11 @@
+/**
+ * @license
+ * Copyright 2023 Google LLC
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+const someSymbol = Symbol();
+
+export class ClassWithUnsupportedProperty {
+  [someSymbol] = 1;
+}