+++ packages/benchmarks/CHANGELOG.md
@@ -11,6 +11,12 @@
   - lit-html@3.0.0-pre.0
   - lit-element@4.0.0-pre.0
 
+## 1.0.3
+
+### Patch Changes
+
+- [#3862](https://github.com/lit/lit/pull/3862) [`a2177c0d`](https://github.com/lit/lit/commit/a2177c0d18d5c6b35b1c4def6911a802a68d828c) - Fixing a typo in the comment
+
 ## 1.0.2
 
 ### Patch Changes