+++ packages/labs/analyzer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lit-labs/analyzer
 
+## 0.7.0
+
+### Minor Changes
+
+- [#3812](https://github.com/lit/lit/pull/3812) [`2c59ceb9`](https://github.com/lit/lit/commit/2c59ceb9427ca76a591084258eedab76644f2a63) - Add CSS Custom Property fallback (default) values to manifest
+
 ## 0.6.1-pre.0
 
 ### Patch Changes