+++ packages/labs/cli/CHANGELOG.md
@@ -11,6 +11,13 @@
   - @lit-labs/gen-utils@0.2.2-pre.0
   - @lit-labs/analyzer@0.6.1-pre.0
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies [[`2c59ceb9`](https://github.com/lit/lit/commit/2c59ceb9427ca76a591084258eedab76644f2a63)]:
+  - @lit-labs/analyzer@0.7.0
+
 ## 0.4.0
 
 ### Minor Changes